Add IRobotEntity base class to braava
