+++ homeassistant/components/roomba/binary_sensor.py
@@ -5,6 +5,7 @@ from homeassistant.components.binary_sensor import BinarySensorDevice
 
 from . import roomba_reported_state
 from .const import BLID, DOMAIN, ROOMBA_SESSION
+from .irobot_base import IRobotEntity
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -21,23 +22,10 @@ async def async_setup_entry(hass, config_entry, async_add_entities):
         async_add_entities([roomba_vac], True)
 
 
-class RoombaBinStatus(BinarySensorDevice):
+class RoombaBinStatus(IRobotEntity, BinarySensorDevice):
     """Class to hold Roomba Sensor basic info."""
 
     ICON = "mdi:delete-variant"
-
-    def __init__(self, roomba, blid):
-        """Initialize the sensor object."""
-        self.vacuum = roomba
-        vacuum_state = roomba_reported_state(roomba)
-        self._blid = blid
-        self._name = vacuum_state.get("name")
-        self._identifier = f"roomba_{self._blid}"
-
-    @property
-    def should_poll(self):
-        """Disable polling."""
-        return False
 
     @property
     def name(self):
@@ -62,19 +50,3 @@ class RoombaBinStatus(BinarySensorDevice):
         )
         _LOGGER.debug("Update Full Bin status from the vacuum: %s", bin_status)
         return bin_status
-
-    @property
-    def device_info(self):
-        """Return the device info of the vacuum cleaner."""
-        return {
-            "identifiers": {(DOMAIN, self._identifier)},
-            "name": str(self._name),
-        }
-
-    def register_callback(self):
-        """Register callback function."""
-        self.vacuum.register_on_message_callback(self.on_message)
-
-    def on_message(self, json_data):
-        """Update state on message change."""
-        self.schedule_update_ha_state()
